+++ JS/exp.js
@@ -40,7 +40,7 @@ tabelElement.addEventListener('click', removeItemFromCart);
 //   };
 // }
 
-
+let counterConfirm = 0;
 function removeItemFromCart(event)
 {
   tabelElement.deleteRow(1);
@@ -108,13 +108,16 @@ function addNewExperiance(event)
   confirmWarning.addEventListener('submit',warningConfirm);
   function warningConfirm()
   {
-
-    let confirmm = confirm('Are You Sure ?');
-    if (confirmm === true) {
-      { let data= JSON.stringify(newExperiance);
-        localStorage.setItem('Experiance',data);}
-    } else {
-      alert('please reconsidering your inputs');
+    counterConfirm++;
+    if(counterConfirm === 1 )
+    {
+      let confirmm = confirm('Are You Sure ?');
+      if (confirmm === true) {
+        { let data= JSON.stringify(newExperiance);
+          localStorage.setItem('Experiance',data);}
+      } else {
+        alert('please reconsidering your inputs');
+      }
     }
   }
 }
